Rearrange packages in persistence
The test packages did not match the implementation structure
