+++ persistence/src/test/kotlin/org/metplus/cruncher/persistence/model/job/JobRepositoryMongoImplTest.kt
@@ -1,6 +1,8 @@
-package org.metplus.cruncher.job
+package org.metplus.cruncher.persistence.model.job
 
 import org.junit.jupiter.api.extension.ExtendWith
+import org.metplus.cruncher.job.JobRepositoryTest
+import org.metplus.cruncher.job.JobsRepository
 import org.metplus.cruncher.persistence.TestConfig
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest
