+++ persistence/src/test/kotlin/org/metplus/cruncher/persistence/model/resume/ResumeRepositoryImplTest.kt
@@ -1,7 +1,9 @@
-package org.metplus.cruncher.settings
+package org.metplus.cruncher.persistence.model.resume
 
 import org.junit.jupiter.api.extension.ExtendWith
 import org.metplus.cruncher.persistence.TestConfig
+import org.metplus.cruncher.resume.ResumeRepository
+import org.metplus.cruncher.resume.ResumeRepositoryTest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.data.mongo.DataMongoTest
 import org.springframework.test.context.ContextConfiguration
@@ -10,11 +12,11 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
 @DataMongoTest
 @ExtendWith(SpringExtension::class)
 @ContextConfiguration(classes = [TestConfig::class])
-class SettingsMongoRepositoryImplTest : SettingsRepositoryTest() {
+internal class ResumeRepositoryImplTest : ResumeRepositoryTest() {
     @Autowired
-    lateinit var mongoRepository: SettingsRepository
+    lateinit var mongoRepository: ResumeRepository
 
-    override fun getRepository(): SettingsRepository {
+    override fun getRepository(): ResumeRepository {
         return mongoRepository
     }
 }
